Create EvenNumbr5.java
here i wrote the even numberchecker by using loops i used both logics in same code here, in real time implementation we have to use for loop or while loop any one of them to do here ,

Using a while loop:
A while loop is used to repeatedly subtract 2 from the number until it becomes zero. If the result is zero, it's even.

Using a for loop:
Similar to the previous example, it uses a for loop to subtract 2 from the number and checks if it becomes zero.